+++ internal/xsql/conn.go
@@ -325,21 +325,3 @@ func (c *conn) BeginTx(ctx context.Context, txOptions driver.TxOptions) (_ drive
 	}
 	return c.currentTx, nil
 }
-
-func Unwrap(db *sql.DB) (connector *Connector, err error) {
-	// hop with create session (connector.Connect()) helps to get ydb.Connection
-	c, err := db.Conn(context.Background())
-	if err != nil {
-		return nil, xerrors.WithStackTrace(err)
-	}
-	if err = c.Raw(func(driverConn interface{}) error {
-		if cc, ok := driverConn.(*conn); ok {
-			connector = cc.connector
-			return nil
-		}
-		return xerrors.WithStackTrace(badconn.Map(fmt.Errorf("%+v is not a *conn", driverConn)))
-	}); err != nil {
-		return nil, xerrors.WithStackTrace(err)
-	}
-	return connector, nil
-}