+++ retry/errors_go1.18.go
@@ -3,6 +3,15 @@
 
 package retry
 
+import (
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
+)
+
 func unwrapErrBadConn(err error) error {
+	var e *badconn.Error
+	if xerrors.As(err, &e) {
+		return e.Origin()
+	}
 	return err
 }